\name{GappedRanges-class}
\docType{class}

% Classes:
\alias{class:GappedRanges}
\alias{GappedRanges-class}
\alias{GappedRanges}

% Generics and methods:
\alias{length,GappedRanges-method}
\alias{start,GappedRanges-method}
\alias{end,GappedRanges-method}
\alias{ngap}
\alias{ngap,GappedRanges-method}
\alias{names,GappedRanges-method}
\alias{names<-,GappedRanges-method}
\alias{coerce,CompressedNormalIRangesList,GappedRanges-method}
\alias{coerce,CompressedIRangesList,GappedRanges-method}
\alias{coerce,GappedRanges,CompressedNormalIRangesList-method}
\alias{coerce,GappedRanges,NormalIRangesList-method}
\alias{coerce,GappedRanges,CompressedIRangesList-method}
\alias{coerce,GappedRanges,IRangesList-method}
\alias{coerce,GappedRanges,RangesList-method}
\alias{as.data.frame,GappedRanges-method}
\alias{show,GappedRanges-method}
\alias{elementNROWS,GappedRanges-method}
\alias{c,GappedRanges-method}


\title{GappedRanges objects (DEFUNCT!)}

\description{
  GappedRanges objects are defunct!
}

\keyword{methods}
\keyword{classes}
