\name{IntegerRanges-class}
\docType{class}

% Classes:
\alias{class:IntegerRanges}
\alias{IntegerRanges-class}
\alias{IntegerRanges}

\title{IntegerRanges objects}

\description{
  The IntegerRanges \emph{virtual} class is a general container for storing
  ranges on the space of integers.
}

\details{
  TODO
}

\keyword{methods}
\keyword{classes}
