\name{IRanges internals}
\docType{class}

\alias{setValidity2}
\alias{new2}
\alias{sapplyLength}

% externalptr methods:
\alias{show,externalptr-method}

% IntegerPtr class and methods:
\alias{!=,IntegerPtr,IntegerPtr-method}
\alias{==,IntegerPtr,IntegerPtr-method}
\alias{[,IntegerPtr-method}
\alias{[<-,IntegerPtr-method}
\alias{as.integer,IntegerPtr-method}
\alias{length,IntegerPtr-method}
\alias{show,IntegerPtr-method}

% NumericPtr class and methods:
\alias{!=,NumericPtr,NumericPtr-method}
\alias{==,NumericPtr,NumericPtr-method}
\alias{[,NumericPtr-method}
\alias{[<-,NumericPtr-method}
\alias{as.numeric,NumericPtr-method}
\alias{length,NumericPtr-method}
\alias{show,NumericPtr-method}


\title{IRanges internals}

\description{
  IRanges objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pages}

\keyword{internal}
\keyword{methods}
\keyword{classes}
