\name{rdapply}
\alias{rdapply}
\alias{rdapply,RDApplyParams-method}
\alias{RDApplyParams-class}

% constructor
\alias{RDApplyParams}

% accessors
\alias{applyFun}
\alias{applyFun,RDApplyParams-method}
\alias{applyFun<-}
\alias{applyFun<-,RDApplyParams-method}
\alias{applyParams}
\alias{applyParams,RDApplyParams-method}
\alias{applyParams<-}
\alias{applyParams<-,RDApplyParams-method}
\alias{filterRules}
\alias{filterRules,RDApplyParams-method}
\alias{filterRules<-}
\alias{filterRules<-,RDApplyParams-method}
\alias{rangedData}
\alias{rangedData,RDApplyParams-method}
\alias{rangedData<-}
\alias{rangedData<-,RDApplyParams-method}
\alias{reducerFun}
\alias{reducerFun,RDApplyParams-method}
\alias{reducerFun<-}
\alias{reducerFun<-,RDApplyParams-method}
\alias{reducerParams}
\alias{reducerParams,RDApplyParams-method}
\alias{reducerParams<-}
\alias{reducerParams<-,RDApplyParams-method}
\alias{simplify}
\alias{simplify,RDApplyParams-method}
\alias{simplify<-}
\alias{simplify<-,RDApplyParams-method}
\alias{iteratorFun}
\alias{iteratorFun,RDApplyParams-method}
\alias{iteratorFun<-}
\alias{iteratorFun<-,RDApplyParams-method}

\title{Applying over spaces (DEFUNCT!)}
\description{
  RDApplyParams objects and the \code{rdapply()} function are defunct!
}

\keyword{classes}
\keyword{methods}
