\name{score}

\alias{score}
\alias{score<-}

\title{Score accessor and setter}

\description{
  Gets and sets the score of an object.
}

\usage{
score(x, ...)
score(x, ...) <- value
}

\arguments{
  \item{x}{An object to get or set the score value of.}
  \item{value}{A new score value.}
  \item{\dots}{Additional arguments.}
}

\keyword{methods}
