\name{XRawList-class}
\docType{class}

\alias{class:XRawList}
\alias{XRawList-class}
\alias{XRawList}


\title{XRawList objects}

\description{
  THIS IS A WORK-IN-PROGRESS!

  An XRawList object is *conceptually* a list of \link{XRaw}
  objects.
}

\author{H. Pages}

\seealso{
  \link{XRaw-class},
  \link{XVectorList-class}
}

\keyword{methods}
\keyword{classes}
