\name{CompressedHitsList-class}
\docType{class}

\alias{class:CompressedHitsList}
\alias{CompressedHitsList-class}
\alias{CompressedHitsList}

% coercion
\alias{as.matrix,CompressedHitsList-method}

% accessors
\alias{space,CompressedHitsList-method}
\alias{from,CompressedHitsList-method}
\alias{to,CompressedHitsList-method}
\alias{nLnode,CompressedHitsList-method}
\alias{nRnode,CompressedHitsList-method}


\title{CompressedHitsList objects}
\description{
  An efficient representation of \link{HitsList} objects.
  See \code{?\link{HitsList}} for more information about \link{HitsList}
  objects.
}

\note{
  This class is highly experimental. It has not been well tested and
  may disappear at any time.
}

\author{Michael Lawrence}

\seealso{
  \link{HitsList} objects.
}

\keyword{methods}
\keyword{classes}
