\name{IRanges internals}
\docType{class}

\alias{setValidity2}
\alias{new2}
\alias{sapplyLength}

% SequencePtr class, functions and methods:
\alias{class:SequencePtr}
\alias{SequencePtr-class}
\alias{SequencePtr}

\alias{show,externalptr-method}
\alias{length,SequencePtr-method}
\alias{==,SequencePtr,SequencePtr-method}
\alias{!=,SequencePtr,SequencePtr-method}

% RawPtr class, functions and methods:
\alias{class:RawPtr}
\alias{RawPtr-class}
\alias{RawPtr}

\alias{RawPtr.readInts}
\alias{RawPtr.writeInts}
\alias{RawPtr.read}
\alias{RawPtr.write}
\alias{RawPtr.copy}
\alias{RawPtr.reverseCopy}
\alias{RawPtr.readComplexes}
\alias{RawPtr.append}
\alias{RawPtr.compare}
\alias{safeExplode}
\alias{[,RawPtr-method}
\alias{[,PrintableRawPtr-method}
\alias{[<-,RawPtr-method}
\alias{as.integer,RawPtr-method}
\alias{show,RawPtr-method}
\alias{show,PrintableRawPtr-method}
\alias{toString,RawPtr-method}

% IntegerPtr class, functions and methods:
\alias{class:IntegerPtr}
\alias{IntegerPtr-class}
\alias{IntegerPtr}

\alias{[,IntegerPtr-method}
\alias{[<-,IntegerPtr-method}
\alias{as.integer,IntegerPtr-method}
\alias{show,IntegerPtr-method}

% NumericPtr class, functions and methods:
\alias{class:NumericPtr}
\alias{NumericPtr-class}
\alias{NumericPtr}

\alias{[,NumericPtr-method}
\alias{[<-,NumericPtr-method}
\alias{as.numeric,NumericPtr-method}
\alias{show,NumericPtr-method}


\title{IRanges internals}

\description{
  IRanges objects, classes and methods that are not intended to
  be used directly.
}

\keyword{internal}
\keyword{methods}
\keyword{classes}
