\name{IRanges internals}

\alias{class:functionORNULL}
\alias{functionORNULL-class}
\alias{functionORNULL}
\alias{class:expressionORfunction}
\alias{expressionORfunction-class}
\alias{coerce,ANY,vector-method}

% For implementation of IRanges extensions
\alias{fixedColumnNames}
\alias{fixedColumnNames,ANY-method}

\title{IRanges internals}

\description{
  Objects, classes and methods defined in the \pkg{IRanges} package
  that are not intended to be used directly.
}

\keyword{internal}
\keyword{classes}
\keyword{methods}
