\name{IRanges internals}
\docType{class}

\alias{setValidity2}
\alias{new2}
\alias{sapplyLength}

% externalptr methods:
\alias{show,externalptr-method}

% RawPtr class, functions and methods:
\alias{class:RawPtr}
\alias{RawPtr-class}
\alias{RawPtr}
\alias{RawPtr.readInts}
\alias{RawPtr.writeInts}
\alias{RawPtr.read}
\alias{RawPtr.write}
\alias{RawPtr.copy}
\alias{RawPtr.reverseCopy}
\alias{RawPtr.readComplexes}
\alias{RawPtr.append}
\alias{RawPtr.compare}
\alias{safeExplode}
\alias{!=,RawPtr,RawPtr-method}
\alias{==,RawPtr,RawPtr-method}
\alias{[,RawPtr-method}
\alias{[,PrintableRawPtr-method}
\alias{[<-,RawPtr-method}
\alias{as.integer,RawPtr-method}
\alias{length,RawPtr-method}
\alias{show,RawPtr-method}
\alias{show,PrintableRawPtr-method}
\alias{toString,RawPtr-method}

% IntegerPtr functions and methods:
\alias{!=,IntegerPtr,IntegerPtr-method}
\alias{==,IntegerPtr,IntegerPtr-method}
\alias{[,IntegerPtr-method}
\alias{[<-,IntegerPtr-method}
\alias{as.integer,IntegerPtr-method}
\alias{length,IntegerPtr-method}
\alias{show,IntegerPtr-method}

% NumericPtr functions and methods:
\alias{!=,NumericPtr,NumericPtr-method}
\alias{==,NumericPtr,NumericPtr-method}
\alias{[,NumericPtr-method}
\alias{[<-,NumericPtr-method}
\alias{as.numeric,NumericPtr-method}
\alias{length,NumericPtr-method}
\alias{show,NumericPtr-method}


\title{IRanges internals}

\description{
  IRanges objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pages}

\keyword{internal}
\keyword{methods}
\keyword{classes}
