\name{IRanges internals}
\docType{class}

\alias{setValidity2}
\alias{new2}
\alias{sapplyLength}

% externalptr methods:
\alias{show,externalptr-method}

% XNumeric class and methods:
\alias{!=,XNumeric,XNumeric-method}
\alias{==,XNumeric,XNumeric-method}
\alias{[,XNumeric-method}
\alias{[<-,XNumeric-method}
\alias{as.numeric,XNumeric-method}
\alias{length,XNumeric-method}
\alias{show,XNumeric-method}


\title{IRanges internals}

\description{
  IRanges objects, classes and methods that are not intended to
  be used directly.
}

\author{H. Pages}

\keyword{internal}
\keyword{methods}
\keyword{classes}
