\name{IRanges internals}

\alias{setValidity2}
\alias{new2}
\alias{setMethods}

\alias{.Call2}
\alias{class:characterORNULL}
\alias{characterORNULL-class}
\alias{characterORNULL}
\alias{class:functionORNULL}
\alias{functionORNULL-class}
\alias{functionORNULL}
\alias{coerce,ANY,vector-method}

% More low-level helper functions (from R/normarg-utils.R):
\alias{isTRUEorFALSE}
\alias{isSingleInteger}
\alias{isSingleNumber}
\alias{isSingleString}
\alias{isSingleNumberOrNA}
\alias{isSingleStringOrNA}
\alias{recycleIntegerArg}
\alias{recycleNumericArg}
\alias{fold}


\title{IRanges internals}

\description{
  Objects, classes and methods defined in the IRanges package that are not
  intended to be used directly.
}

\keyword{internal}
\keyword{methods}
\keyword{classes}
