\name{RangedDataList-class}
\docType{class}
\alias{RangedDataList-class}
\alias{unlist,RangedDataList-method}
\alias{stack,RangedDataList-method}

% Constructor
\alias{RangedDataList}

\title{Lists of RangedData (DEFUNCT!)}
\description{
  \code{RangedDataList} objects are defunct!
}

\keyword{classes}
