\name{XVectorList-class}
\docType{class}

\alias{class:GroupedIRanges}
\alias{GroupedIRanges-class}
\alias{GroupedIRanges}

\alias{class:XVectorList}
\alias{XVectorList-class}
\alias{XVectorList}

\alias{as.data.frame,GroupedIRanges-method}
\alias{show,GroupedIRanges-method}
\alias{[,GroupedIRanges-method}
\alias{seqselect,GroupedIRanges-method}
\alias{window,GroupedIRanges-method}
\alias{c,GroupedIRanges-method}

\alias{length,XVectorList-method}
\alias{width,XVectorList-method}
\alias{names,XVectorList-method}
\alias{names<-,XVectorList-method}
\alias{[[,XVectorList-method}
\alias{[,XVectorList-method}
\alias{seqselect,XVectorList-method}
\alias{window,XVectorList-method}
\alias{subseq,XVectorList-method}
\alias{c,XVectorList-method}
\alias{[<-,XVectorList-method}
\alias{[[<-,XVectorList-method}

\alias{unlist_list_of_XVectorList}
\alias{unsplit_list_of_XVectorList}


\title{XVectorList objects}

\description{
  THIS IS A WORK-IN-PROGRESS!

  An XVectorList object is *conceptually* a list of \link{XVector}
  objects.
}

\author{H. Pages}

\seealso{
  \link{XVector-class},
  \link{XRawList-class},
  \code{\link{compact}}
}

\keyword{methods}
\keyword{classes}
